change location test
