+++ tests/types/interfaces/test_fluctuations_functions.py
@@ -1,8 +1,10 @@
 from athena.core.interfaces.fluctuations import (
     merge_candles,
     convert_candles_to_period,
+    sanitize_candles,
 )
-from athena.core.types import Period
+from athena.core.types import Period, Coin
+from athena.core.interfaces import Candle
 
 import datetime
 
@@ -71,3 +73,47 @@ def test_convert_candles_to_period_wrong_timeframe(generate_candles):
 
     with pytest.raises(ValueError, match="Cannot convert candles to lower timeframe"):
         convert_candles_to_period(candles, target_period)
+
+
+@pytest.fixture
+def raw_candles():
+    other_params = {
+        "open": 100,
+        "close": 100,
+        "high": 150,
+        "low": 70,
+        "coin": Coin.BTC,
+        "currency": Coin.USDT,
+        "period": "1h",
+        "open_time": datetime.datetime(2024, 1, 1),
+        "close_time": datetime.datetime(2024, 1, 2),
+        "quote_volume": 150,
+        "nb_trades": 100,
+        "taker_volume": 50,
+        "taker_quote_volume": 70,
+    }
+
+    def _raw_candles(volumes: list[int]) -> list[Candle]:
+        return [
+            Candle(
+                volume=volume,
+                **other_params,
+            )
+            for volume in volumes
+        ]
+
+    return _raw_candles
+
+
+def test_sanitize_candles(raw_candles):
+    """Remove invalid candles."""
+
+    sanitized_candles = sanitize_candles(raw_candles(volumes=[50, 60, 0, 100, 0]))
+
+    assert [candle.volume for candle in sanitized_candles] == [50, 60, 100]
+
+
+def test_sanitize_candles_empty():
+    sanitized_candles = sanitize_candles([])
+
+    assert sanitized_candles == []
